+++ src/index.ts
@@ -2,7 +2,7 @@ import { DebugContext } from "./Context";
 import {
   DebugContextProps,
   DebugItemProps,
-  InsetDebugItemProps
+  InsetDebugItemProps,
 } from "./Context/types";
 import Inset from "./Inset";
 import insetFactory from "./Inset/insetFactory";
@@ -15,7 +15,7 @@ import {
   PaddingPossibilities,
   VerHor,
   Vertical,
-  InsetStyles
+  InsetStyles,
 } from "./Inset/insetTypes";
 import Stack from "./Stack";
 import stackFactory from "./Stack/stackFactory";
@@ -23,6 +23,13 @@ import { StackDebugOptions, StackProps, StackStyles } from "./Stack/stackTypes";
 import Queue from "./Queue";
 import queueFactory from "./Queue/queueFactory";
 import { QueueDebugOptions, QueueProps, QueueStyles } from "./Queue/queueTypes";
+
+function spacingFactory<T>(spacing: { [K in keyof T]: number }) {
+  const Inset = insetFactory<T>(spacing);
+  const Stack = stackFactory<T>(spacing);
+  const Queue = queueFactory<T>(spacing);
+  return { Inset, Stack, Queue };
+}
 
 export {
   DebugContext,
@@ -40,6 +47,7 @@ export {
   VerHor,
   Vertical,
   InsetStyles,
+  spacingFactory,
   Stack,
   stackFactory,
   StackDebugOptions,
@@ -49,5 +57,5 @@ export {
   queueFactory,
   QueueDebugOptions,
   QueueProps,
-  QueueStyles
+  QueueStyles,
 };